requirements.txt
streamlit
pandas
plotly
openpyxl
streamlit-gsheets-connection